fixing more stuff
