+++ logic/ibr_logic.py
@@ -8,9 +8,9 @@ from data_handling.shared_data import filtrar_por_fecha
 from logic.shared_logic import (
     calcular_fecha_anterior,
     convertir_tasa_nominal_a_efectiva_anual,
-    shift_list_with_replacement,
     sumar_tasas,
 )
+from utils.helper_functions import shift_list_with_replacement
 
 co_holidays = holidays.Colombia()  # Festivos en Colombia
 
@@ -238,13 +238,14 @@ def sumar_spread_ibr_batch(
         raise Exception(f"Error al calcular la tasa de negociación IBR: {str(e)}")
 
 
-def procesar_tasa_cupon_ibr_online(
+def procesar_tasa_cupon_ibr_datos(
     base_dias_anio: str,
     periodicidad: str,
     tasa_anual_cupon: float,
     lista_fechas: list[str],
     fecha_negociacion: datetime.date,
     modalidad: str,
+    archivo,
 ):
     """
     Procesa una tasa Spread Cupon. Método online.
@@ -290,7 +291,10 @@ def procesar_tasa_cupon_ibr_online(
         num_per=1,
     )
     tasa_per_anterior = sumar_spread_ibr(
-        tasa_spread=tasa_anual_cupon, fecha=fecha_per_anterior, modalidad=modalidad
+        tasa_spread=tasa_anual_cupon,
+        fecha=fecha_per_anterior,
+        modalidad=modalidad,
+        archivo=archivo,
     )
 
     # Calcular la tasa para el primer cupón
@@ -299,7 +303,10 @@ def procesar_tasa_cupon_ibr_online(
 
     # Obtener IBR del día anterior a la fecha de negociación para los siguientes cupones
     ibr_negociacion = sumar_spread_ibr(
-        tasa_spread=tasa_anual_cupon, fecha=fecha_negociacion, modalidad=modalidad
+        tasa_spread=tasa_anual_cupon,
+        fecha=fecha_negociacion,
+        modalidad=modalidad,
+        archivo=archivo,
     )
     for _ in range(1, len(lista_fechas)):
         tasa_ibr_spread_i = (ibr_negociacion) / 100
@@ -314,6 +321,7 @@ def procesar_tasa_flujos_real_ibr_online(
     tasa_anual_cupon: float,
     lista_fechas: list[str],
     modalidad: str,
+    archivo,
 ):
     """
     Procesa una tasa Spread Cupon. Método online.
@@ -359,11 +367,17 @@ def procesar_tasa_flujos_real_ibr_online(
         num_per=1,
     )
     tasa_per_anterior = sumar_spread_ibr(
-        tasa_spread=tasa_anual_cupon, fecha=fecha_per_anterior, modalidad=modalidad
+        tasa_spread=tasa_anual_cupon,
+        fecha=fecha_per_anterior,
+        modalidad=modalidad,
+        archivo=archivo,
     )
 
     tasa_fechas = sumar_spread_ibr_batch(
-        tasa_spread=tasa_anual_cupon, lista_fechas=fechas_cupones, modalidad=modalidad
+        tasa_spread=tasa_anual_cupon,
+        lista_fechas=fechas_cupones,
+        modalidad=modalidad,
+        archivo=archivo,
     )
     # reemplazar por la anterior
     tasa_fechas = shift_list_with_replacement(
@@ -379,111 +393,6 @@ def procesar_tasa_flujos_real_ibr_online(
     tasa_fechas = [round(x - tasa_anual_cupon, 2) for x in tasa_fechas]
 
     return tasas_final, tasa_fechas
-
-
-def procesar_tasa_cupon_ibr_proyectado(
-    base_dias_anio: str,
-    periodicidad: str,
-    tasa_anual_cupon: float,
-    lista_fechas: list[str],
-    fecha_inicio: datetime.date,
-    fecha_negociacion: datetime.date,
-    archivo=None,
-):
-    """
-    Convierte una tasa nominal anual a una nominal en otra periodicidad.
-
-    Parámetros:
-    base_dias_anio (str): Base de cálculo de días ('30/360' o '365/365').
-    periodicidad (str): Periodo de conversión ('Mensual', 'Trimestral', 'Semestral', 'Anual').
-    tasa_anual_cupon (float): Tasa anual expresada en decimal (Ej: 10% -> 0.10).
-    lista_fechas (list[str]): Lista de fechas de cada cupón en formato 'DD/MM/YYYY'.
-    fecha_inicio (datetime.date): Fecha de inicio para la consulta de tasas IBR.
-    fecha_negociacion (datetime.date): Fecha de negociación en formato 'DD/MM/YYYY'.
-    archivo (str, opcional): Nombre del archivo Excel que contiene las tasas IBR.
-
-    Retorna:
-    list[float]: Lista de tasas convertidas a la periodicidad especificada.
-    """
-    tasa_anual_cupon = tasa_anual_cupon / 100
-    base = {"30/360": 360, "365/365": 365}
-    periodos_por_anio = {"Mensual": 12, "Trimestral": 4, "Semestral": 2, "Anual": 1}
-
-    if not lista_fechas:
-        raise ValueError("La lista de fechas de cupones está vacía.")
-
-    if periodicidad not in periodos_por_anio:
-        raise ValueError(
-            "Periodicidad no válida. Usa: 'Mensual', 'Trimestral', 'Semestral' o 'Anual'."
-        )
-
-    if base_dias_anio not in base:
-        raise ValueError("Base no válida. Usa '30/360' o '365/365'.")
-
-    # Convertir lista de fechas a objetos datetime y calcular fechas reales de IBR
-    fechas_cupones = [
-        datetime.datetime.strptime(f, "%d/%m/%Y").date() for f in lista_fechas
-    ]
-    fechas_reales_ibr = [fecha_publicacion_ibr(f) for f in fechas_cupones]
-
-    # Obtener tasas IBR en batch
-    tasas_ibr = filtrar_por_fecha(
-        archivo=archivo, nombre_hoja="IBR Estimada", fechas_filtro=fechas_reales_ibr
-    )
-
-    # Asegurar que tasas_ibr es una serie o lista antes de dividir
-    if isinstance(tasas_ibr, pd.DataFrame):
-        if tasas_ibr.empty:
-            raise ValueError(
-                "No se encontraron datos IBR en el rango de fechas especificado."
-            )
-    # Convert list to a DataFrame
-    df_fechas_reales_ibr = pd.DataFrame({"Fecha": fechas_reales_ibr})
-    # Convert Fecha column to datetime format
-    df_fechas_reales_ibr["Fecha"] = pd.to_datetime(df_fechas_reales_ibr["Fecha"])
-    df_merged = df_fechas_reales_ibr.merge(tasas_ibr, on="Fecha", how="left")
-    # Convertir a diccionario asegurando que la estructura es correcta
-    tasas_ibr_dict = dict(zip(df_merged.iloc[:, 0].dt.date, df_merged.iloc[:, 1]))
-
-    # definir tasas
-    tasas = []
-    # Obtener IBR del día anterior a la fecha anterior del cupon proximo a vencer
-    fecha_per_anterior = calcular_fecha_anterior(
-        fecha=min(fechas_cupones),
-        periodicidad=periodicidad,
-        base_intereses=base_dias_anio,
-        num_per=1,
-    )
-    fecha_per_anterior_real = fecha_publicacion_ibr(fecha_per_anterior)
-    ibr_per_anterior_real = filtrar_por_fecha(
-        archivo=archivo,
-        nombre_hoja="IBR Estimada",
-        fechas_filtro=[fecha_per_anterior_real],
-    )
-    if ibr_per_anterior_real.empty:
-        raise ValueError(
-            f"No se encontraron datos IBR en la fecha del periodo anterior ({fecha_per_anterior_real}) del cupon proximo a vencer."
-        )
-
-    # Calcular la tasa para el primer cupón
-    tasa_ibr_spread_1 = ibr_per_anterior_real.iloc[0, 1] + tasa_anual_cupon
-    tasas.append(tasa_ibr_spread_1 / periodos_por_anio[periodicidad])
-
-    # Obtener IBR del día anterior a la fecha de negociación para los siguientes cupones
-    for i in range(1, len(lista_fechas)):
-        fecha_real_ibr_i = fechas_reales_ibr[i]
-        tasa_ibr_i = tasas_ibr_dict.get(fecha_real_ibr_i, None)
-
-        if tasa_ibr_i is None or pd.isna(tasa_ibr_i):
-            raise ValueError(
-                f"""No se encontró la tasa IBR para la fecha: {fecha_real_ibr_i}.
-                Por favor usa la opcion de subir un archivo con las tasas proyectadss"""
-            )
-
-        tasa_ibr_spread_i = tasa_ibr_i + tasa_anual_cupon
-        tasas.append(tasa_ibr_spread_i / periodos_por_anio[periodicidad])
-
-    return tasas
 
 
 def es_dia_habil_bancario(fecha: datetime.date) -> bool:
